+++ internal/metrics/prometheus.go
@@ -9,19 +9,19 @@ import (
 	ctrlmetrics "sigs.k8s.io/controller-runtime/pkg/metrics"
 )
 
-// Recorder is the interface fir recording metrics on a certain operation.
+// Recorder is the interface for recording metrics for provided operation.
 type Recorder interface {
-	// REVIEW: add URL of Konnect API as a label?
-	RecordKonnectEntityOperationSuccess(operationType KonnectEntityOperation, entityType string, duration time.Duration)
-	RecordKonnectEntityOperationFailure(operationType KonnectEntityOperation, entityType string, duration time.Duration, statusCode int)
+	RecordKonnectEntityOperationSuccess(serverURL string, operationType KonnectEntityOperation, entityType string, duration time.Duration)
+	RecordKonnectEntityOperationFailure(serverURL string, operationType KonnectEntityOperation, entityType string, duration time.Duration, statusCode int)
 }
 
 type KonnectEntityOperation string
 
 const (
+	KonnectServerURLKey                                  = "server_url"
 	KonnectEntityOperationTypeKey                        = "operation_type"
 	KonnectEntityOperationCreate  KonnectEntityOperation = "create"
-	KonnectEnttiyOperationUpdate  KonnectEntityOperation = "update"
+	KonnectEntityOperationUpdate  KonnectEntityOperation = "update"
 	KonnectEntityOperationDelete  KonnectEntityOperation = "delete"
 
 	KonnectEntityTypeKey = "entity_type"
@@ -35,9 +35,8 @@ const (
 
 // metric names for konnect entity operations.
 const (
-	// REVIEW: define a Namespace `gateway_operator` for creating prometheus metrics here?
 	MetricNameKonnectEntityOperationCount    = "gateway_operator_konnect_entity_operation_count"
-	MetricNameKonnectEntityOperationDuration = "gateway_operator_konnect_entity_operation_duration"
+	MetricNameKonnectEntityOperationDuration = "gateway_operator_konnect_entity_operation_duration_milliseconds"
 )
 
 var (
@@ -46,37 +45,45 @@ var (
 			Name: MetricNameKonnectEntityOperationCount,
 			Help: fmt.Sprintf(
 				"Count of successful/failed entity operations in Konnect. "+
+					"`%s` describes the URL of the Konnect server. "+
 					"`%s` describes the operation type (`%s`, `%s`, or `%s`)."+
 					"`%s` describes the type of the operated entity. "+
 					"`%s` describes whether the operation is successful (`%s`) or not (`%s`). "+
-					"`%s` is populated in case of `%s=\"%s\"` and describes the status code returned from Konnect API.",
-				KonnectEntityOperationTypeKey, KonnectEntityOperationCreate, KonnectEnttiyOperationUpdate, KonnectEntityOperationDelete,
+					"`%s` is populated in case of `%s=\"%s\"` and describes the status code returned from Konnect API. "+
+					"`%s`=\"0\" and %s=\"%s\" means we cannot collect the status code or error happens in the process other than Konnect API call.",
+				KonnectServerURLKey,
+				KonnectEntityOperationTypeKey, KonnectEntityOperationCreate, KonnectEntityOperationUpdate, KonnectEntityOperationDelete,
 				KonnectEntityTypeKey,
 				SuccessKey, SuccessTrue, SuccessFalse,
 				StatusCodeKey, SuccessKey, SuccessFalse,
+				StatusCodeKey, SuccessKey, SuccessFalse,
 			),
 		},
-		[]string{KonnectEntityOperationTypeKey, KonnectEntityTypeKey, SuccessKey, StatusCodeKey},
+		[]string{KonnectServerURLKey, KonnectEntityOperationTypeKey, KonnectEntityTypeKey, SuccessKey, StatusCodeKey},
 	)
 
 	konnectEntityOperationDuration = prometheus.NewHistogramVec(
 		prometheus.HistogramOpts{
 			Name: MetricNameKonnectEntityOperationDuration,
 			Help: fmt.Sprintf(
-				"How long did the Konnect entity operation take in seconds. "+
+				"How long did the Konnect entity operation take in milliseconds. "+
+					"`%s` describes the URL of the Konnect server. "+
 					"`%s` describes the operation type (`%s`, `%s`, or `%s`)."+
 					"`%s` describes the type of the operated entity. "+
 					"`%s` describes whether the operation is successful (`%s`) or not (`%s`). "+
-					"`%s` is populated in case of `%s=\"%s\"` and describes the status code returned from Konnect API.",
-				KonnectEntityOperationTypeKey, KonnectEntityOperationCreate, KonnectEnttiyOperationUpdate, KonnectEntityOperationDelete,
+					"`%s` is populated in case of `%s=\"%s\"` and describes the status code returned from Konnect API. "+
+					"`%s`=\"0\" and %s=\"%s\" means we cannot collect the status code or error happens in the process other than Konnect API call.",
+				KonnectServerURLKey,
+				KonnectEntityOperationTypeKey, KonnectEntityOperationCreate, KonnectEntityOperationUpdate, KonnectEntityOperationDelete,
 				KonnectEntityTypeKey,
 				SuccessKey, SuccessTrue, SuccessFalse,
 				StatusCodeKey, SuccessKey, SuccessFalse,
+				StatusCodeKey, SuccessKey, SuccessFalse,
 			),
-			// Duration range from 0.1s to 10min (600s).
-			Buckets: prometheus.ExponentialBucketsRange(0.1, 600, 20),
+			// Duration range from 1ms to 10min.
+			Buckets: prometheus.ExponentialBucketsRange(1, 10*float64(time.Minute.Milliseconds()), 20),
 		},
-		[]string{KonnectEntityOperationTypeKey, KonnectEntityTypeKey, SuccessKey, StatusCodeKey},
+		[]string{KonnectServerURLKey, KonnectEntityOperationTypeKey, KonnectEntityTypeKey, SuccessKey, StatusCodeKey},
 	)
 )
 
@@ -95,34 +102,36 @@ func NewGlobalCtrlRuntimeMetricsRecorder() *GlobalCtrlRuntimeMetricsRecorder {
 }
 
 func (r *GlobalCtrlRuntimeMetricsRecorder) RecordKonnectEntityOperationSuccess(
-	operationType KonnectEntityOperation, entityType string, duration time.Duration) {
-	r.recordKonnectEntityOperationCount(operationType, entityType, true, 0)
-	r.recordKonnectEntityOperationDuration(operationType, entityType, true, 0, duration)
+	serverURL string, operationType KonnectEntityOperation, entityType string, duration time.Duration) {
+	r.recordKonnectEntityOperationCount(serverURL, operationType, entityType, true, 0)
+	r.recordKonnectEntityOperationDuration(serverURL, operationType, entityType, true, 0, duration)
 }
 
 func (r *GlobalCtrlRuntimeMetricsRecorder) RecordKonnectEntityOperationFailure(
-	operationType KonnectEntityOperation, entityType string, duration time.Duration, statusCode int) {
-	r.recordKonnectEntityOperationCount(operationType, entityType, false, statusCode)
-	r.recordKonnectEntityOperationDuration(operationType, entityType, false, statusCode, duration)
+	serverURL string, operationType KonnectEntityOperation, entityType string, duration time.Duration, statusCode int) {
+	r.recordKonnectEntityOperationCount(serverURL, operationType, entityType, false, statusCode)
+	r.recordKonnectEntityOperationDuration(serverURL, operationType, entityType, false, statusCode, duration)
 }
 
 func (r *GlobalCtrlRuntimeMetricsRecorder) recordKonnectEntityOperationCount(
-	operationType KonnectEntityOperation, entityType string, success bool, statusCode int,
+	serverURL string, operationType KonnectEntityOperation, entityType string, success bool, statusCode int,
 ) {
-	labels := konnectEntityOperationLabels(operationType, entityType, success, statusCode)
+	labels := konnectEntityOperationLabels(serverURL, operationType, entityType, success, statusCode)
 	konnectEntityOperationCount.With(labels).Inc()
 }
 
 func (r *GlobalCtrlRuntimeMetricsRecorder) recordKonnectEntityOperationDuration(
-	operationType KonnectEntityOperation, entityType string, success bool, statusCode int, duration time.Duration,
+	serverURL string, operationType KonnectEntityOperation, entityType string, success bool, statusCode int, duration time.Duration,
 ) {
-	labels := konnectEntityOperationLabels(operationType, entityType, success, statusCode)
+	labels := konnectEntityOperationLabels(serverURL, operationType, entityType, success, statusCode)
 	konnectEntityOperationDuration.With(labels).Observe(duration.Seconds())
 }
 
-func konnectEntityOperationLabels(operationType KonnectEntityOperation, entityType string, success bool, statusCode int,
+func konnectEntityOperationLabels(
+	serverURL string, operationType KonnectEntityOperation, entityType string, success bool, statusCode int,
 ) prometheus.Labels {
 	labels := prometheus.Labels{
+		KonnectServerURLKey:           serverURL,
 		KonnectEntityOperationTypeKey: string(operationType),
 		KonnectEntityTypeKey:          entityType,
 	}
@@ -146,14 +155,16 @@ func init() {
 	}
 }
 
+// TODO: move all the mocks to a place inside `/test`:
+// https://github.com/Kong/gateway-operator/issues/955
 type MockRecorder struct{}
 
 var _ Recorder = &MockRecorder{}
 
 func (m *MockRecorder) RecordKonnectEntityOperationSuccess(
-	operationType KonnectEntityOperation, entityType string, duration time.Duration) {
+	serverURL string, operationType KonnectEntityOperation, entityType string, duration time.Duration) {
 }
 
 func (m *MockRecorder) RecordKonnectEntityOperationFailure(
-	operationType KonnectEntityOperation, entityType string, duration time.Duration, statusCode int) {
+	serverURL string, operationType KonnectEntityOperation, entityType string, duration time.Duration, statusCode int) {
 }